requirements.txt
pytorch>=1.0
numpy>=1.14
gensim==3.7.3
scipy>=
networkx>=2.1
tqdm==4.24.0
scikit-learn==0.19.1

